fix type error for duration. isolate compute duration.
