+++ frontend/src/app/courses/page.tsx
@@ -1,15 +1,20 @@
-import { Prisma } from "@prisma/client";
 import Link from "next/link";
+import { type Prisma } from "@prisma/client";
+import { computeCourseDuration } from "@/lib/utils";
 
 type Course = Prisma.CourseGetPayload<{
     include: { instructor: true };
-}>;
+}> & { duration: number };
 
 export const dynamic = "force-dynamic";
 
 export default async function Courses() {
     const res = await fetch(process.env.BACKEND_API_URL + "/courses");
     const courses: Course[] = await res.json();
+
+    for (const course of courses) {
+        computeCourseDuration(course);
+    }
 
     return (
         <div className="grid grid-cols-1 md:grid-cols-3 lg:grid-cols-5 gap-6">
